Fixes to match current OLS data.
